[DEV-9731] Update file names
