+++ packages/chart/index.html
@@ -101,7 +101,7 @@
     <!-- <div class="react-container" data-config="/https://www.cdc.gov/poxvirus/mpox/modules/data-viz/mpx-age-sex1.json"></div> -->
 
     <!-- TESTS DATE EXCLUSIONS -->
-    <div class="react-container" data-config="/examples/feature/boxplot/boxplot.json"></div>
+    <!-- <div class="react-container" data-config="/examples/feature/boxplot/boxplot.json"></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/tests-case-rate/case-rate-example-config.json"></div> -->
 
     <!-- TESTS BIG SMALL-->
