+++ packages/chart/src/_stories/Chart.Legend.Gradient.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import chartGradientConfig from './_mock/legend.gradient_mock.json'
 import SimplifiedLineConfig from './_mock/simplified_line.json'
 
-import Chart from '../CdcChartComponent'
+import Chart from '../CdcChart'
 import { editConfigKeys } from '../helpers/configHelpers'
 
 const meta: Meta<typeof Chart> = {
