+++ packages/chart/src/_stories/Chart.ScatterPlot.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Chart from '../CdcChartComponent'
+import Chart from '../CdcChart'
 import scatterPlotDownloadImage from './_mock/scatterplot-image-download.json'
 
 const meta: Meta<typeof Chart> = {
