+++ packages/chart/src/_stories/Chart.tooltip.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Chart from '../CdcChartComponent'
+import Chart from '../CdcChart'
 
 const meta: Meta<typeof Chart> = {
   title: 'Components/Templates/Chart/Tooltip',
