+++ packages/dashboard/src/CdcDashboardComponent.tsx
@@ -19,7 +19,7 @@ import Loading from '@cdc/core/components/Loading'
 import { DataTransform } from '@cdc/core/helpers/DataTransform'
 import getViewport from '@cdc/core/helpers/getViewport'
 
-import CdcChartComponent from '@cdc/chart/src/CdcChartComponent'
+import CdcChart from '@cdc/chart/src/CdcChart'
 import CdcDataBite from '@cdc/data-bite/src/CdcDataBite'
 import CdcMap from '@cdc/map/src/CdcMap'
 import CdcWaffleChart from '@cdc/waffle-chart/src/CdcWaffleChart'
@@ -352,7 +352,7 @@ export default function CdcDashboard({ initialState, isEditor = false, isDebug =
             body = (
               <>
                 <Header visualizationKey={visualizationKey} subEditor='Chart' />
-                <CdcChartComponent
+                <CdcChart
                   key={visualizationKey}
                   config={visualizationConfig}
                   isEditor={true}
